feat: add an -f (--fix) flag to actually fix files
This commit adds laravel/prompts to allow the user to fix issues by
choosing from a SuggestPrompt where they can see the suggested
spellings or choose their own.

There's also a -s (--first-suggestiong) flag that will automatically
choose the first suggestion for each issue.

The testing is a bit janky because I ran into some issues with
Symphony's testing and trying to 'fake' the -s cli flag, so instead I
temporarily copied one of the testing files to the ./src folder, run
peck on itself alowing it to fix that file, and then compare with
expected content before cleaning up the ./src folder.